create Context folder and file with initial creation of context
